third_party/WebKit/LayoutTests/editing/execCommand/outdent-blockquote-test4-expected.txt

CONSOLE MESSAGE: line 12: SUCCESS: Nodes stayed in order after outdent.
This test uses the execCommand to Outdent the text below. This test that nodes, with different relationships with blockquotes, stay in the correct order after an outdent.



A

B
